+++ basic/src/bin/valida.rs
@@ -1,13 +1,9 @@
-use byteorder::{LittleEndian, ReadBytesExt};
 use clap::Parser;
-use std::fs::File;
-use std::io::{BufReader, Read, Result, Write};
+use std::io::{Read, Write};
 
 use valida_basic::BasicMachine;
 use valida_cpu::MachineWithCpuChip;
-use valida_machine::{
-    InstructionWord, Machine, Operands, ProgramROM, Word, MEMORY_CELL_BYTES, OPERAND_ELEMENTS,
-};
+use valida_machine::{Machine, ProgramROM, Word, MEMORY_CELL_BYTES};
 use valida_output::MachineWithOutputChip;
 use valida_program::MachineWithProgramChip;
 
@@ -28,7 +24,7 @@ fn main() {
     let args = Args::parse();
 
     let mut machine = BasicMachine::<BabyBear, BabyBear>::default();
-    let rom = load_program_rom(&args.program).unwrap();
+    let rom = ProgramROM::from_file(&args.program).unwrap();
     machine.program_mut().set_program_rom(&rom);
     machine.cpu_mut().fp = args.stack_height;
     machine.cpu_mut().save_register_state();
@@ -65,21 +61,4 @@ fn main() {
                 .collect::<Vec<_>>(),
         )
         .unwrap();
-}
-
-fn load_program_rom(filename: &str) -> Result<ProgramROM<i32>> {
-    let file = File::open(filename)?;
-    let mut reader = BufReader::new(file);
-    let mut instructions = Vec::new();
-
-    while let Ok(opcode) = reader.read_u32::<LittleEndian>() {
-        let mut operands_arr = [0i32; OPERAND_ELEMENTS];
-        for i in 0..OPERAND_ELEMENTS {
-            operands_arr[i] = reader.read_i32::<LittleEndian>()?;
-        }
-        let operands = Operands(operands_arr);
-        instructions.push(InstructionWord { opcode, operands });
-    }
-
-    Ok(ProgramROM::new(instructions))
 }
