+++ machine/Cargo.toml
@@ -4,6 +4,10 @@ version = "0.1.0"
 edition = "2021"
 license = "MIT OR Apache-2.0"
 
+[features]
+default = ["std"]
+std = []
+
 [dependencies]
 byteorder = "1.4.3"
 itertools = "0.10.3"
